Move colorscheme hash into helper function and refactor node position functions
