+++ src/renderer/helpers/colorSchemeHash.tsx
